fix: IdTextField 공통 컴포넌트 생성
