+++ app/src/main/java/org/sopt/at/core/component/textfield/IdTextField.kt
@@ -1,0 +1,48 @@
+package org.sopt.at.core.component.textfield
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.unit.dp
+import org.sopt.at.R
+
+@Composable
+fun IdTextField(
+    id: String,
+    onIdChange: (String) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    TextField(
+        value = id,
+        onValueChange = onIdChange,
+        placeholder = { Text(text = stringResource(R.string.textfield_id)) },
+        modifier = modifier
+            .fillMaxWidth(),
+        shape = RoundedCornerShape(5.dp),
+        colors = TextFieldDefaults.colors(
+            focusedContainerColor = Color(0xFF262626),
+            unfocusedContainerColor = Color(0xFF262626),
+            focusedIndicatorColor = Color.Transparent,
+            unfocusedIndicatorColor = Color.Transparent,
+            disabledIndicatorColor = Color.Transparent,
+            cursorColor = Color.White,
+            focusedPlaceholderColor = Color.Gray,
+            unfocusedPlaceholderColor = Color.Gray
+        ),
+        textStyle = TextStyle(
+            color = Color.White
+        ),
+        singleLine = true,
+        keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next)
+    )
+}
